added de-registration of chainnode upon shutdown
it's a bit of a hack - the uri assigned by the directory node is stored in the config class.
Works if the chain node is shut down in by stopping mvn:exec using ctrl-c, but does not if the
chain node process is killed using KILL. In order to maintain an up-to-date list, the directory
node should poll the chain nodes in regular intervals.